# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: FactorSage API
  version: "1.0.0"
  summary: Stock research, strategy design, backtesting and monitoring.
  description: |
    The HTTP API behind FactorSage.

    **Authentication** is a session cookie issued by `POST /auth/login` or by the Google OAuth
    callback. It is `HttpOnly`, so browser clients send it automatically and never read it; a
    non-browser client must store and resend it. There is no bearer token and no API key.

    **Authorization** has two independent axes, both resolved server-side from persisted state and
    never from anything the client sends: a role (`USER` / `ADMIN`) and a commercial plan
    (`FREE` / `STARTER` / `PRO`). A plan limit is refused with `403` and a stable
    `ENTITLEMENT_*` code. Ownership is not a separate axis — every user-owned collection is scoped
    to the caller, and somebody else's resource answers exactly like one that does not exist.

    **Rate limiting** is applied per endpoint class from a central policy catalog, keyed by user id
    for authenticated traffic and by client IP otherwise. Exceeding a policy answers `429` with
    `Retry-After`. See `info.x-rate-limit-policies`.

    **Errors** share one machine-readable shape: an HTTP `statusCode`, a human `message`, and — for
    every refusal a client may reasonably branch on — a stable `code`. Clients should branch on
    `code`, never on `message`.
  contact:
    name: FactorSage engineering
  license:
    name: UNLICENSED

...

...
servers:
...
paths:
...
  /strategies:
    get:
      tags: [Strategies]
      operationId: listStrategies
      summary: The caller's strategies.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: standard-read }
      responses:
        "200":
          description: Summaries, without the definitions.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StrategySummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

    post:
      tags: [Strategies]
      operationId: createStrategy
      summary: Create a strategy.
      description: |
        The definition is validated by the same canonical validator the Strategy Builder runs, so a
        rejection carries path-addressed `issues` the UI attaches to the exact row that produced
        them rather than showing a banner.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: mutation }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateStrategyRequest" }
      responses:
        "201":
          description: The created strategy, with its canonical normalized definition.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StrategyDetail" }
        "400": { $ref: "#/components/responses/StrategyInvalid" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/EntitlementDenied" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /strategies/{strategyId}:
    parameters:
      - { $ref: "#/components/parameters/StrategyId" }
    get:
      tags: [Strategies]
      operationId: getStrategy
      summary: One strategy with its full definition.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: standard-read }
      responses:
        "200":
          description: The strategy.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StrategyDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/OwnedResourceNotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

    patch:
      tags: [Strategies]
      operationId: updateStrategy
      summary: Rename a strategy or change its description.
      description: Metadata only. The definition is replaced through its own endpoint.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: mutation }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateStrategyRequest" }
      responses:
        "200":
          description: The updated summary.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StrategySummary" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/OwnedResourceNotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

    delete:
      tags: [Strategies]
      operationId: deleteStrategy
      summary: Delete a strategy.
      description: |
        Completed backtest runs are unaffected: a run holds an immutable snapshot of the strategy it
        executed, so deleting the live strategy never deletes or reinterprets a result.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: mutation }
      responses:
        "204":
          description: Deleted.
          headers: *rateLimitHeaders
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/OwnedResourceNotFound" }
        "409":
          description: A monitor still references this strategy. Delete or rebind the monitor first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /strategies/{strategyId}/definition:
    parameters:
      - { $ref: "#/components/parameters/StrategyId" }
    put:
      tags: [Strategies]
      operationId: replaceStrategyDefinition
      summary: Replace the complete strategy definition.
      description: |
        Atomic whole-document replace; it creates a new strategy version. Partial edits are not
        supported — send the entire definition.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: mutation }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              { $ref: "#/components/schemas/ReplaceStrategyDefinitionRequest" }
      responses:
        "200":
          description: The strategy with its new canonical definition and incremented version.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StrategyDetail" }
        "400": { $ref: "#/components/responses/StrategyInvalid" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/OwnedResourceNotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ---------------------------------------------------------------------------------------------
  # Backtests
  # ---------------------------------------------------------------------------------------------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: intrinsic_auth
      description: |
        The session cookie issued by `POST /auth/login` or by the Google callback. `HttpOnly`,
        `SameSite=Lax`, and `Secure` in production. Its name is deployment configuration
        (`AUTH_COOKIE_NAME`); `intrinsic_auth` is the default. A browser sends it automatically —
        cross-origin callers must send credentials explicitly, and the deployment's CORS allowlist
        must contain their origin.
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: |
        Signature over the raw request body, verified with the webhook signing secret. The only
        authentication the webhook endpoint has, and the only one it needs.
  headers:
    RateLimitPolicy:
      description: The rate-limit policy that produced the numbers in the other headers.
      schema: { type: string }
      example: standard-read
    RateLimitLimit:
      description: The policy's allowance for the current window.
      schema: { type: integer, minimum: 1 }
      example: 240
    RateLimitRemaining:
      description: Allowance left after this request.
      schema: { type: integer, minimum: 0 }
      example: 239
    RateLimitReset:
      description: Whole seconds until the allowance refills.
      schema: { type: integer, minimum: 1 }
      example: 57
    RetryAfter:
      description: |
        Whole seconds to wait before retrying. Always at least 1 — a `Retry-After: 0` would invite
        the immediate retry that caused the refusal. On `429` it is measured from the counter's own
        remaining window, so waiting exactly this long is sufficient.
      schema: { type: integer, minimum: 1 }
      example: 42
  parameters:
...
    StrategyId:
      name: strategyId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No session, or a session that is expired, forged or otherwise unusable.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    EntitlementDenied:
      description: |
        The caller's plan does not allow this. The body carries a stable `ENTITLEMENT_*` code plus
        the numbers behind it, so a client can say "your plan allows 10, this change would make 11"
        without parsing the message.

        This is a **product limit**, not a rate limit: retrying identically will fail identically,
        and the remedy is a different plan or a smaller request — never waiting.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/EntitlementError" }
    OwnedResourceNotFound:
      description: |
        No such resource for this caller. A resource that exists but belongs to somebody else
        answers exactly like one that does not exist, so this response reveals nothing about another
        user's data.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    StrategyInvalid:
      description: |
        The strategy definition was rejected. Carries `code: STRATEGY_INVALID` and path-addressed
        `issues`, each naming the level, condition and field that produced it.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/StrategyValidationError" }
...
    TooManyRequests:
      description: |
        The caller exceeded this endpoint's rate-limit policy. `Retry-After` is the measured time
        until the allowance refills; waiting exactly that long is sufficient, and retrying sooner is
        refused again without consuming anything further.

        A client should surface this as throttling with the wait, never as a generic failure — and
        on `POST /auth/login` in particular it must not be shown as "wrong credentials", because the
        attempt is refused whether or not the password was correct.
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
        RateLimit-Policy: { $ref: "#/components/headers/RateLimitPolicy" }
        RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
        RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
        RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/RateLimitError" }
          examples:
            search:
              value:
                statusCode: 429
                message: "Too many requests. Retry in 42 seconds."
                code: RATE_LIMITED
                policy: stock-search
                retryAfterSeconds: 42
...
  schemas:
...
    Error:
      type: object
      description: |
        The API's error shape. `code` is present on every refusal a client can reasonably branch on
        and absent on generic framework errors; branch on it, never on `message`.
      required: [statusCode, message]
      properties:
        statusCode: { type: integer, description: Repeats the HTTP status. }
        message:
          {
            type: string,
            description: Human-readable. Never parsed by a client.,
          }
        code:
          {
            type: string,
            description: Stable machine-readable reason,
            when there is one.,
          }
      examples:
        - { statusCode: 404, message: "Stock symbol was not found" }
    RateLimitError:
      allOf:
        - { $ref: "#/components/schemas/Error" }
        - type: object
          required: [code, policy, retryAfterSeconds]
          properties:
            code:
              type: string
              enum: [RATE_LIMITED, RATE_LIMIT_UNAVAILABLE]
              description: |
                `RATE_LIMITED` — the caller exceeded an allowance (`429`).
                `RATE_LIMIT_UNAVAILABLE` — the limiter could not answer and this endpoint's policy
                is fail-closed (`503`). The caller's allowance is untouched.
            policy:
              type: string
              description: |
                Which class of endpoint refused, so a client can tell "your typing is too fast" from
                "you have submitted too many backtests" without matching on the path. Stable
                identifier; the numbers behind it are operational configuration.
              enum:
                - auth-sensitive
                - session-probe
                - stock-search
                - stock-read
                - standard-read
                - progress-poll
                - mutation
                - monitor-mutation
                - backtest-execution
                - billing-mutation
                - admin-operation
            retryAfterSeconds:
              type: integer
              minimum: 1
              description: Mirrors `Retry-After`. Whole seconds, rounded up, never zero.
    EntitlementError:
      allOf:
        - { $ref: "#/components/schemas/Error" }
        - type: object
          required: [code, tier]
          properties:
            code:
              type: string
              enum:
                - ENTITLEMENT_AUTH_REQUIRED
                - ENTITLEMENT_FEATURE_UNAVAILABLE
                - ENTITLEMENT_LIST_SYMBOL_LIMIT
                - ENTITLEMENT_BACKTEST_SYMBOL_LIMIT
                - ENTITLEMENT_BACKTEST_HISTORY_LIMIT
                - ENTITLEMENT_BACKTEST_CONCURRENCY_LIMIT
                - ENTITLEMENT_MONITOR_LIMIT
                - ENTITLEMENT_RESOURCE_OVER_LIMIT
            tier: { $ref: "#/components/schemas/EntitlementTier" }
            limit:
              { type: integer, description: What the caller's plan allows. }
            current: { type: integer, description: What they already have. }
            requested:
              { type: integer, description: What the request would make it. }
...
    StrategyValidationError:
      type: object
      required: [message, code, issues]
      properties:
        message: { type: string }
        code: { type: string, const: STRATEGY_INVALID }
        issues:
          type: array
          items: { $ref: "#/components/schemas/StrategyValidationIssue" }
    StrategyValidationIssue:
      type: object
      required: [code, path, message]
      properties:
        code:
          type: string
          enum:
            [
              SHAPE_INVALID,
              UNKNOWN_FIELD,
              NAME_REQUIRED,
              NAME_TOO_LONG,
              DESCRIPTION_TOO_LONG,
              BUY_LEVEL_REQUIRED,
              TOO_MANY_LEVELS,
              TOO_MANY_CONDITIONS,
              SIGNAL_EMPTY,
              PERCENTAGE_INVALID,
              METRIC_NOT_ALLOWED_IN_LEVEL,
              METRIC_SERIES_UNSUPPORTED,
              OPERATOR_NOT_SUPPORTED,
              VALUE_KIND_MISMATCH,
              VALUE_OUT_OF_DOMAIN,
              SERIES_UNKNOWN,
              SERIES_NOT_COMPARABLE,
              DUPLICATE_CONDITION,
              DUPLICATE_ID,
            ]
        message: { type: string }
        path:
          type: object
          description: Addresses the exact element that failed, so a builder can attach the issue to it.
          required: [part]
          properties:
            levelKind: { $ref: "#/components/schemas/StrategyLevelKind" }
            levelIndex: { type: integer, minimum: 0 }
            part:
              type: string
              enum:
                [
                  STRATEGY,
                  NAME,
                  DESCRIPTION,
                  LEVEL,
                  PERCENTAGE,
                  CONDITION,
                  TRIGGER,
                ]
            conditionIndex: { type: integer, minimum: 0 }
            field: { type: string, enum: [METRIC, OPERATOR, VALUE] }

    # -------------------------------------------------------------------------------------------
    # Identity and access
    # -------------------------------------------------------------------------------------------
...
    EntitlementTier:
      type: string
      enum: [GUEST, FREE, STARTER, PRO, ADMIN]
      description: Which entitlement set was applied. Diagnostics and UI only.
...
    StrategyLevelKind:
      type: string
      enum: [BUY, SELL, FINAL_EXIT]
    SelectableSeriesId:
      type: string
      description: |
        An identifier from the canonical selectable-series catalog — a moving average, an oscillator,
        an intrinsic-value model or a blend. One catalog serves Stock Details and strategy
        conditions alike, so there is no second list of options anywhere.
      examples: [sma50d, rsi14d, iv-blend-balanced]
    StrategyMetric:
      description: The left-hand side of a condition or a trigger.
      oneOf:
        - type: object
          required: [kind]
          properties: { kind: { type: string, const: PRICE } }
        - type: object
          required: [kind, seriesId]
          properties:
            kind: { type: string, const: MOVING_AVERAGE }
            seriesId: { $ref: "#/components/schemas/SelectableSeriesId" }
        - type: object
          required: [kind, seriesId]
          properties:
            kind: { type: string, const: OSCILLATOR }
            seriesId: { $ref: "#/components/schemas/SelectableSeriesId" }
        - type: object
          required: [kind, sourceId]
          description: Margin of safety against one intrinsic-value series. Percent domain, at most 100.
          properties:
            kind: { type: string, const: MARGIN_OF_SAFETY }
            sourceId: { $ref: "#/components/schemas/SelectableSeriesId" }
        - type: object
          description: |
            Position-dependent, so it is only meaningful where simulated position state exists;
            allowed in SELL and FINAL_EXIT levels, never in BUY.
          required: [kind]
          properties: { kind: { type: string, const: GAIN } }
        - type: object
          required: [kind]
          properties: { kind: { type: string, const: LOSS } }
    StrategyValue:
      description: |
        The right-hand side. Domains are metric-specific: margin of safety is at most 100, gain is
        at least -100, loss is 0..100, and a series value must be comparable with its metric — same
        timeframe, and never the metric's own series.
      oneOf:
        - type: object
          required: [kind, seriesId]
          properties:
            kind: { type: string, const: SERIES }
            seriesId: { $ref: "#/components/schemas/SelectableSeriesId" }
        - type: object
          required: [kind, value]
          properties:
            kind: { type: string, const: NUMBER }
            value: { type: number }
        - type: object
          required: [kind, value]
          properties:
            kind: { type: string, const: PERCENT }
            value: { type: number }
    StrategyCondition:
      type: object
      required: [id, metric, operator, value]
      properties:
        id: { type: string }
        metric: { $ref: "#/components/schemas/StrategyMetric" }
        operator: { type: string, enum: [IS_ABOVE, IS_BELOW, IS_CLOSE_TO] }
        value: { $ref: "#/components/schemas/StrategyValue" }
    StrategyTrigger:
      type: object
      description: |
        An event on one observation date rather than a state, which is what makes a trigger level
        fire at most once per date.
      required: [id, metric, operator, value]
      properties:
        id: { type: string }
        metric: { $ref: "#/components/schemas/StrategyMetric" }
        operator: { type: string, enum: [CROSSES_ABOVE, CROSSES_BELOW] }
        value: { $ref: "#/components/schemas/StrategyValue" }
    StrategySignal:
      type: object
      description: |
        Zero or more conditions, ANDed, plus at most one optional trigger which is ANDed with them
        for the same date. A signal with neither is empty and is rejected.
      required: [conditions]
      properties:
        conditions:
          type: array
          maxItems: 10
          items: { $ref: "#/components/schemas/StrategyCondition" }
        trigger: { $ref: "#/components/schemas/StrategyTrigger" }
    StrategyDefinition:
      type: object
      description: |
        The complete strategy. At least one BUY level is required. A level owns exactly one signal
        and one percentage; there is no stock list, no valuation source and no execution parameter
        here — capital, contributions, period and `maximumPositions` all belong to a backtest.
      required: [schemaVersion, buyLevels, sellLevels]
      properties:
        schemaVersion:
          { type: integer, description: Definition schema version. }
        buyLevels:
          type: array
          minItems: 1
          maxItems: 10
          items:
            type: object
            required: [id, signal, percentage]
            properties:
              id: { type: string }
              signal: { $ref: "#/components/schemas/StrategySignal" }
              percentage:
                type: integer
                enum: [25, 50, 75, 100]
                description: Fraction of a full position this level buys.
        sellLevels:
          type: array
          maxItems: 10
          items:
            type: object
            required: [id, signal, percentage]
            properties:
              id: { type: string }
              signal: { $ref: "#/components/schemas/StrategySignal" }
              percentage: { type: integer, enum: [25, 50, 75] }
        finalExit:
          type: object
          description: Closes the whole remaining position when its signal matches.
          required: [id, signal]
          properties:
            id: { type: string }
            signal: { $ref: "#/components/schemas/StrategySignal" }
    StrategySummary:
      type: object
      required:
        [
          id,
          name,
          buyLevelCount,
          sellLevelCount,
          hasFinalExit,
          versionNumber,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 500 }
        buyLevelCount: { type: integer, minimum: 0 }
        sellLevelCount: { type: integer, minimum: 0 }
        hasFinalExit: { type: boolean }
        versionNumber:
          {
            type: integer,
            minimum: 1,
            description: Incremented by every definition replace.,
          }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    StrategyDetail:
      allOf:
        - { $ref: "#/components/schemas/StrategySummary" }
        - type: object
          required: [definition]
          properties:
            definition: { $ref: "#/components/schemas/StrategyDefinition" }
    CreateStrategyRequest:
      type: object
      required: [name, definition]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, maxLength: 500 }
        definition: { $ref: "#/components/schemas/StrategyDefinition" }
    UpdateStrategyRequest:
      type: object
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description:
          oneOf:
            - { type: string, maxLength: 500 }
            - { type: "null" }
    ReplaceStrategyDefinitionRequest:
      type: object
      required: [definition]
      properties:
        definition: { $ref: "#/components/schemas/StrategyDefinition" }

    # -------------------------------------------------------------------------------------------
    # Benchmarks
    # -------------------------------------------------------------------------------------------
...